Added locales and documentation for FakeDataVariableParsingHHandler
