--- src/HttPlaceholder.Application/Files/Markdown/fake_person_data-description.md
@@ -1,1 +1,0 @@
-TODO
